tests/wpt/tests/html/canvas/element/fill-and-stroke-styles/2d.fillStyle.colorObject.transparency.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.fillStyle.colorObject.transparency</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.fillStyle.colorObject.transparency</h1>
<p class="desc">ctx.fillStyle with color objects has transparency</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("ctx.fillStyle with color objects has transparency");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = {r: 0, g: 1, b: 0, a: 0};
  ctx.fillRect(0, 0, 100, 50);
  _assertPixel(canvas, 50,25, 0,0,0,0);
  ctx.clearRect(0, 0, 100, 50);
  ctx.fillStyle = {r: 0, g: 1, b: 0, a: -1};
  ctx.fillRect(0, 0, 100, 50);
  _assertPixel(canvas, 50,25, 0,0,0,0);
  ctx.clearRect(0, 0, 100, 50);
  ctx.fillStyle = {r: 0, g: 1, b: 0, a: 0.5};
  ctx.fillRect(0, 0, 100, 50);
  _assertPixel(canvas, 50,25, 0,255,0,128);
  ctx.clearRect(0, 0, 100, 50);
  ctx.fillStyle = {r: 0, g: 1, b: 0, a: 1};
  ctx.fillRect(0, 0, 100, 50);
  _assertPixel(canvas, 50,25, 0,255,0,255);

});
</script>
